+++ src/main/java/itson/ticketwizard/entidades/TransaccionCompra.java
@@ -1,5 +1,6 @@
 package itson.ticketwizard.entidades;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
@@ -13,12 +14,12 @@ public class TransaccionCompra extends Transaccion {
     public TransaccionCompra() {
     }
 
-    public TransaccionCompra(Integer codigoTransaccion, Integer codigoUsuario, LocalDateTime fechaHoraTransaccion, double monto, String numeroSerieBoleto) {
+    public TransaccionCompra(Integer codigoTransaccion, Integer codigoUsuario, LocalDateTime fechaHoraTransaccion, BigDecimal monto, String numeroSerieBoleto) {
         super(codigoTransaccion, codigoUsuario, fechaHoraTransaccion, monto, numeroSerieBoleto);
         this.codigoTransaccion = codigoTransaccion;
     }
 
-    public TransaccionCompra(Integer codigoUsuario, LocalDateTime fechaHoraTransaccion, double monto, String numeroSerieBoleto) {
+    public TransaccionCompra(Integer codigoUsuario, LocalDateTime fechaHoraTransaccion, BigDecimal monto, String numeroSerieBoleto) {
         super(codigoUsuario, fechaHoraTransaccion, monto, numeroSerieBoleto);
     }
